Remote-Device-Management/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-ads1x15==2.4.1
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.10.1
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
asyncio==3.4.3
awscrt==0.21.1
awsiotsdk==1.22.0
binho-host-adapter==0.1.6
boto3==1.33.13
botocore==1.33.13
certifi==2024.12.14
charset-normalizer==3.4.0
distlib==0.3.9
distro==1.9.0
filelock==3.12.2
future==1.0.0
idna==3.10
importlib-metadata==6.7.0
iso8601==2.1.0
jmespath==1.0.1
numpy==2.2.0
pandas==2.2.3
platformdirs==4.0.0
pyftdi==0.54.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.2
pyusb==1.2.1
PyYAML==6.0.1
requests==2.31.0
RPLCD==1.3.1
s3transfer==0.8.2
schedule==1.2.2
scipy==1.14.1
serial==0.0.97
six==1.16.0
smbus2==0.5.0
sysv_ipc==1.1.0
tabulate==0.9.0
typing_extensions==4.7.1
tzdata==2025.1
urllib3==1.26.20
utils==1.0.2
zipp==3.15.0
